Rename `modifying` package to `gumtree`
