+++ plugin/src/main/kotlin/org/jetbrains/research/plugin/gumtree/PyPsiGumTree.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.plugin.modifying
+package org.jetbrains.research.plugin.gumtree
 
 import com.github.gumtreediff.tree.ITree
 import com.github.gumtreediff.tree.Tree
